feature: prepared moving PocoBuilder to common
